Extract Analytics types
